recipe info feature preview

+++ feature/feed/src/main/java/lithium/kotlin/recipesbook/feature/feed/navigation/FeedScreenNavigation.kt
@@ -14,7 +14,7 @@ fun NavController.navigateToFeedScreen(navOptions: NavOptions)
     = this.navigate(FEED_ROUTE, navOptions)
 
 fun NavGraphBuilder.feedScreen(
-    onRecipeClick: () -> Unit
+    onRecipeClick: (Long) -> Unit
 ) {
     composable(
         route = FEED_ROUTE,
